requirements.txt
Flask
PyPDF2
python-docx
Pillow
pytesseract
pymupdf
transformers
